dashboard/backend/requirements.txt
# Updated dependencies for dashboard backend
fastapi>=0.109.0  # Updated from 0.104.1
uvicorn[standard]>=0.27.0 # Updated from 0.24.0
python-dotenv>=1.0.1
pydantic>=2.5.3  # Updated from 2.5.0
pydantic-settings>=2.1.0 # Should be compatible with pydantic 2.x
supabase>=2.4.0  # Updated from 2.3.5 - check for latest stable
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
# Note: 'cors' is not a standard package name. FastAPI handles CORS via middleware.
# If you need CORS, add 'python-multipart' is already listed, or add specific version if needed.
# Example for CORS middleware usage in FastAPI app (not a package to install separately)
# from fastapi.middleware.cors import CORSMiddleware
# You likely don't need a 'cors' package in requirements.txt
# Remove 'cors==1.0.1' unless it refers to a specific, non-standard package you intend to use.